+++ src/models/deviceSettings/shutterSettings.ts
@@ -5,6 +5,11 @@ export class ShutterSettings extends DeviceSettings {
   public msTilTop: number = -1;
   public msTilBot: number = -1;
   /**
+   * The direction this shutter is facing (0 = North, 180 = South)
+   * @type {number}
+   */
+  public direction?: number;
+  /**
    * Some shutter give no position feedback on their own, so by knowing the durations in either direction,
    * we can programmatically trigger the callbacks.
    * @type {boolean}
@@ -14,6 +19,7 @@ export class ShutterSettings extends DeviceSettings {
   public fromPartialObject(data: Partial<ShutterSettings>): void {
     this.msTilTop = data.msTilTop ?? this.msTilTop;
     this.msTilBot = data.msTilBot ?? this.msTilBot;
+    this.direction = data.direction ?? this.direction;
     this.triggerPositionUpdateByTime = data.triggerPositionUpdateByTime ?? this.triggerPositionUpdateByTime;
     super.fromPartialObject(data);
   }