+++ src/server/devices/baseDeviceInterfaces/iShutter.ts
@@ -1,7 +1,9 @@
 import { Window } from '../groups';
 import { iRoomDevice } from './iRoomDevice';
+import { ShutterSettings } from '../../../models';
 
 export interface iShutter extends iRoomDevice {
+  settings: ShutterSettings;
   currentLevel: number;
   desiredWindowShutterLevel: number;
   window: Window | undefined;