+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Move direction setting from Window to the Shutter, to prevent group-settings need
+
 ## 2.9.11 (2023-06-05)
 
 * (theimo1221) Deep Omit should compare case insensetive in regards to keys to omit